Moved `SoundHolder` class to another file
