+++ app/src/main/java/com/bignerdranch/android/beatbox/MainActivity.kt
@@ -30,24 +30,6 @@ class MainActivity : AppCompatActivity() {
         super.onDestroy()
     }
 
-    private inner class SoundHolder(
-        private val binding: ListItemSoundBinding,
-        beatBox: BeatBox
-        ):
-            RecyclerView.ViewHolder(binding.root) {
-
-        init {
-            binding.viewModel = SoundViewModel(beatBox)
-        }
-
-        fun bind(sound: Sound) {
-            binding.apply {
-                viewModel?.sound = sound
-                executePendingBindings()
-            }
-        }
-    }
-
     private inner class SoundAdapter(
         private val beatBox: BeatBox
         ):
